Engine: renamed AudioCoreSlot into AudioPlayer and picked out
