+++ Engine/media/audio/audio_core.cpp
@@ -19,6 +19,7 @@
 #include <thread>
 #include <unordered_map>
 #include "debug/out.h"
+#include "media/audio/audioplayer.h"
 #include "media/audio/sdldecoder.h"
 #include "media/audio/openalsource.h"
 #include "util/memory_compat.h"
@@ -26,173 +27,11 @@
 using namespace AGS::Common;
 using namespace AGS::Engine;
 
-const auto GlobalGainScaling = 0.7f; // TODO: find out why 0.7f is here?
-
-static void audio_core_entry();
-
-// AudioCoreSlot is a single playback manager, that handles two components:
-// decoder and "player"; controls the current playback state, passes data
-// from the decoder into the player.
-class AudioCoreSlot
-{
-public:
-    AudioCoreSlot(int handle, std::unique_ptr<SDLDecoder> decoder);
-
-    // Gets current playback state
-    PlaybackState GetPlayState() const { return _playState; }
-    // Gets duration, in ms
-    float GetDurationMs() const { return _decoder->GetDurationMs(); }
-    // Gets playback position, in ms
-    float GetPositionMs() const { return _source->GetPositionMs(); }
-    // Gives access to decoder object
-    SDLDecoder &GetDecoder() const { return *_decoder; }
-    // Gives access to the "player" object
-    OpenAlSource &GetAlSource() const { return *_source; }
-
-    // Update state, transfer data from decoder to player if possible
-    void Poll();
-    // Begin playback
-    void Play();
-    // Pause playback
-    void Pause();
-    // Stop playback completely
-    void Stop();
-    // Seek to the given time position
-    void Seek(float pos_ms);
-
-private:
-    // Opens decoder and sets up playback state
-    void Init();
-
-    const int handle_ = -1; // for diagnostic purposes only
-    std::unique_ptr<SDLDecoder> _decoder;
-    std::unique_ptr<OpenAlSource> _source;
-    PlaybackState _playState = PlayStateInitial;
-    PlaybackState _onLoadPlayState = PlayStatePaused;
-    float _onLoadPositionMs = 0.0f;
-    SoundBuffer _bufferPending{};
-};
-
-AudioCoreSlot::AudioCoreSlot(int handle, std::unique_ptr<SDLDecoder> decoder)
-    : handle_(handle), _decoder(std::move(decoder))
-{
-    _source = std::make_unique<OpenAlSource>(
-        _decoder->GetFormat(), _decoder->GetChannels(), _decoder->GetFreq());
-}
-
-void AudioCoreSlot::Init()
-{
-    bool success;
-    if (_decoder->IsValid()) // if already opened, then just seek to start
-        success = _decoder->Seek(_onLoadPositionMs) == _onLoadPositionMs;
-    else
-        success = _decoder->Open(_onLoadPositionMs);
-    _playState = success ? _onLoadPlayState : PlayStateError;
-    if (_playState == PlayStatePlaying)
-        _source->Play();
-}
-
-void AudioCoreSlot::Poll()
-{
-    if (_playState == PlaybackState::PlayStateInitial)
-        Init();
-    if (_playState != PlayStatePlaying)
-        return;
-
-    // Read data from Decoder and pass into the Al Source
-    if (!_bufferPending.Data && !_decoder->EOS())
-    { // if no buffer saved, and still something to decode, then read a buffer
-        _bufferPending = _decoder->GetData();
-        assert(_bufferPending.Data || (_bufferPending.Size == 0));
-    }
-    if (_bufferPending.Data && (_bufferPending.Size > 0))
-    { // if having a buffer already, then try to put into source
-        if (_source->PutData(_bufferPending) > 0)
-            _bufferPending = SoundBuffer(); // clear buffer on success
-    }
-    _source->Poll();
-    // If both finished decoding and playing, we done here.
-    if (_decoder->EOS() && _source->IsEmpty())
-    {
-        _playState = PlayStateFinished;
-    }
-}
-
-void AudioCoreSlot::Play()
-{
-    switch (_playState)
-    {
-    case PlayStateInitial:
-        _onLoadPlayState = PlayStatePlaying;
-        break;
-    case PlayStateStopped:
-        _decoder->Seek(0.0f);
-        /* fall-through */
-    case PlayStatePaused:
-        _playState = PlayStatePlaying;
-        _source->Play();
-        break;
-    default:
-        break;
-    }
-}
-
-void AudioCoreSlot::Pause()
-{
-    switch (_playState)
-    {
-    case PlayStateInitial:
-        _onLoadPlayState = PlayStatePaused;
-        break;
-    case PlayStatePlaying:
-        _playState = PlayStatePaused;
-        _source->Pause();
-        break;
-    default:
-        break;
-    }
-}
-
-void AudioCoreSlot::Stop()
-{
-    switch (_playState)
-    {
-    case PlayStateInitial:
-        _onLoadPlayState = PlayStateStopped;
-        break;
-    case PlayStatePlaying:
-    case PlayStatePaused:
-        _playState = PlayStateStopped;
-        _source->Stop();
-        _bufferPending = SoundBuffer(); // clear
-        break;
-    default:
-        break;
-    }
-}
-
-void AudioCoreSlot::Seek(float pos_ms)
-{
-    switch (_playState)
-    {
-    case PlayStateInitial:
-        _onLoadPositionMs = pos_ms;
-        break;
-    case PlayStatePlaying:
-    case PlayStatePaused:
-    case PlayStateStopped:
-        {
-            _source->Stop();
-            _bufferPending = SoundBuffer(); // clear
-            float new_pos = _decoder->Seek(pos_ms);
-            _source->SetPlaybackPosMs(new_pos);
-        }
-        break;
-    default:
-        break;
-    }
-}
-
+// A volume scale factor between AGS->OpenAL
+// TODO: find out why 0.7f is here?
+// TODO: should move this to OpenAL output?
+// but one problem is that this is also used for the "master volume".
+const auto GlobalGainScaling = 0.7f;
 
 // Global audio core state and resources
 static struct 
@@ -214,7 +53,7 @@ static struct
     // polled one by one, any action like pause/resume is also synced.
     std::mutex mixer_mutex_m;
     std::condition_variable mixer_cv;
-    std::unordered_map<int, std::unique_ptr<AudioCoreSlot>> slots_;
+    std::unordered_map<int, std::unique_ptr<AudioPlayer>> slots_;
 } g_acore;
 
 // Prints any OpenAL errors to the log
@@ -229,6 +68,8 @@ void dump_al_errors()
 // -------------------------------------------------------------------------------------------------
 // INIT / SHUTDOWN
 // -------------------------------------------------------------------------------------------------
+
+static void audio_core_entry();
 
 void audio_core_init() 
 {
@@ -308,7 +149,7 @@ static int audio_core_slot_init(std::unique_ptr<SDLDecoder> decoder)
 {
     auto handle = avail_slot_id();
     std::lock_guard<std::mutex> lk(g_acore.mixer_mutex_m);
-    g_acore.slots_[handle] = std::make_unique<AudioCoreSlot>(handle, std::move(decoder));
+    g_acore.slots_[handle] = std::make_unique<AudioPlayer>(handle, std::move(decoder));
     g_acore.mixer_cv.notify_all();
     return handle;
 }
@@ -380,10 +221,10 @@ void audio_core_set_master_volume(float newvol)
 void audio_core_slot_configure(int slot_handle, float volume, float speed, float panning)
 {
     std::lock_guard<std::mutex> lk(g_acore.mixer_mutex_m);
-    auto &player = g_acore.slots_[slot_handle]->GetAlSource();
-    player.SetVolume(volume * GlobalGainScaling);
-    player.SetSpeed(speed);
-    player.SetPanning(panning);
+    auto *player = g_acore.slots_[slot_handle].get();
+    player->SetVolume(volume * GlobalGainScaling);
+    player->SetSpeed(speed);
+    player->SetPanning(panning);
 }
 
 // -------------------------------------------------------------------------------------------------
@@ -393,7 +234,7 @@ void audio_core_slot_configure(int slot_handle, float volume, float speed, float
 float audio_core_slot_get_pos_ms(int slot_handle)
 {
     std::lock_guard<std::mutex> lk(g_acore.mixer_mutex_m);
-    auto pos = g_acore.slots_[slot_handle]->GetAlSource().GetPositionMs();
+    auto pos = g_acore.slots_[slot_handle]->GetPositionMs();
     g_acore.mixer_cv.notify_all();
     return pos;
 }
@@ -401,7 +242,7 @@ float audio_core_slot_get_pos_ms(int slot_handle)
 float audio_core_slot_get_duration(int slot_handle)
 {
     std::lock_guard<std::mutex> lk(g_acore.mixer_mutex_m);
-    auto dur = g_acore.slots_[slot_handle]->GetDecoder().GetDurationMs();
+    auto dur = g_acore.slots_[slot_handle]->GetDurationMs();
     g_acore.mixer_cv.notify_all();
     return dur;
 }
@@ -409,7 +250,7 @@ float audio_core_slot_get_duration(int slot_handle)
 int audio_core_slot_get_freq(int slot_handle)
 {
     std::lock_guard<std::mutex> lk(g_acore.mixer_mutex_m);
-    auto dur = g_acore.slots_[slot_handle]->GetDecoder().GetFreq();
+    auto dur = g_acore.slots_[slot_handle]->GetFrequency();
     g_acore.mixer_cv.notify_all();
     return dur;
 }
@@ -426,7 +267,7 @@ PlaybackState audio_core_slot_get_play_state(int slot_handle, float &pos_ms)
 {
     std::lock_guard<std::mutex> lk(g_acore.mixer_mutex_m);
     auto state = g_acore.slots_[slot_handle]->GetPlayState();
-    pos_ms = g_acore.slots_[slot_handle]->GetAlSource().GetPositionMs();
+    pos_ms = g_acore.slots_[slot_handle]->GetPositionMs();
     g_acore.mixer_cv.notify_all();
     return state;
 }
